requirements.txt
# Core libraries
pandas>=2.0.0
numpy>=1.25.0

# LLM and RAG
llama-index>=0.7.30
openai>=0.28.0

# Vector search (if using FAISS)
faiss-cpu>=1.7.4

# FastAPI server
fastapi>=0.109.0
uvicorn[standard]>=0.23.0

# Optional: for re-ranking and hybrid search
scikit-learn>=1.3.0

# Environment management (optional)
python-dotenv>=1.0.0

